pyproject.toml
[project]
name = "neuro-tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.21.0",
    "colorlog>=6.9.0",
    "dotenv>=0.9.9",
    "openai>=1.98.0",
    "pyyaml>=6.0.2",
]
